+++ src/com/continuousx/jenkins/features/AbstractFeature.groovy
@@ -2,8 +2,8 @@ package com.continuousx.jenkins.features
 
 import com.cloudbees.groovy.cps.NonCPS
 import com.continuousx.jenkins.LogLevelType
-import com.continuousx.jenkins.features.github.GitURLParser
-import com.continuousx.jenkins.features.jenkins.utils.JenkinsPluginCheck
+import com.continuousx.utils.github.GitURLParser
+import com.continuousx.utils.jenkins.JenkinsPluginCheck
 import com.continuousx.jenkins.features.metrics.influxdb.InfluxDBFeature
 import com.continuousx.jenkins.features.metrics.influxdb.InfluxDBFeatureBuilder
 import com.continuousx.jenkins.features.metrics.influxdb.measurements.operating.MeasurementOperatingFeature
