+++ src/com/continuousx/jenkins/stages/AbstractStage.groovy
@@ -1,7 +1,7 @@
 package com.continuousx.jenkins.stages
 
-import com.continuousx.jenkins.features.github.GitURLParser
-import com.continuousx.jenkins.features.jenkins.utils.JenkinsPluginCheck
+import com.continuousx.utils.github.GitURLParser
+import com.continuousx.utils.jenkins.JenkinsPluginCheck
 import com.continuousx.jenkins.features.metrics.influxdb.InfluxDBFeature
 import com.continuousx.jenkins.features.metrics.influxdb.InfluxDBFeatureBuilder
 import com.continuousx.jenkins.features.metrics.influxdb.measurements.operating.MeasurementOperatingPipelineStage
