+++ src/com/continuousx/jenkins/stages/StageConfig.groovy
@@ -1,13 +1,14 @@
 package com.continuousx.jenkins.stages
 
 import com.continuousx.jenkins.LogLevelType
-import com.continuousx.jenkins.features.github.BranchType
+import com.continuousx.utils.github.BranchType
 
 interface StageConfig extends Serializable {
 
     StageType getType()
     boolean isActive()
     boolean isFailOnError()
+
     BranchType getAllowedBranch()
     LogLevelType logLevelType()
 }
