+++ src/com/continuousx/utils/github/BranchExpressionChecker.groovy
@@ -1,4 +1,4 @@
-package com.continuousx.jenkins.features.github
+package com.continuousx.utils.github
 
 class BranchExpressionChecker {
