+++ test/com/continuousx/jenkins/pipelines/utils/JenkinsPluginCheckTest.groovy
@@ -1,7 +1,7 @@
 package com.continuousx.jenkins.pipelines.utils
 
 
-import com.continuousx.jenkins.features.jenkins.utils.JenkinsPluginCheck
+import com.continuousx.utils.jenkins.JenkinsPluginCheck
 import org.codehaus.groovy.runtime.powerassert.PowerAssertionError
 import spock.lang.Specification
